feat(cli): allow executing plans from plan workflow
This introduces the ability for users to directly execute a generated plan from within the 'plan' workflow.

A new 'Accept and Execute' option has been added to the prompt after a plan is generated. Selecting this option will immediately start the execution of the plan as a task.

To enable this, a new `runTask` tool was added, which leverages the existing `Runner` to execute the provided task. The plan workflow has been updated to call this new tool.