feat(cli): make output silent by default
Most commands will no longer stream tool events to the console by default. The `meta` and `code` commands are exceptions and will continue to show tool events.

To see tool events for any command, use the `-v` flag. To see more detailed output, use `-vv`.

This change introduces a default verbosity level of -1, which silences the event stream. The `meta` and `code` commands default to a verbosity level of 0.